+++ bin/index.js
@@ -11,8 +11,6 @@ const chalk = require('chalk')
 
 const help = require('../lib/help.js')
 const pkg = require('../package.json')
-const BlinkMobileIdentity = require('@blinkmobile/bm-identity')
-const blinkMobileIdentity = new BlinkMobileIdentity(pkg.name)
 
 // this module
 
@@ -27,21 +25,36 @@ const cli = meow({
   help,
   version: true
 }, {
-  boolean: [ 'prune', 'skip', 'debug' ],
+  boolean: [
+    'debug',
+    'force',
+    'prune',
+    'skip'
+  ],
   default: {
+    debug: false,
+    cwd: process.cwd(),
+    force: false,
+    env: 'dev',
     skip: true
   },
-  string: ['bucket', 'region']
+  string: [
+    'bucket',
+    'cwd',
+    'env',
+    'region'
+  ]
 })
 
 const command = cli.input[0]
 
 if (!command) {
-  cli.showHelp(1)
+  cli.showHelp(0)
 }
 
 if (!commands[command]) {
-  console.error(`unknown command: ${command}`)
+  console.error(chalk.red(`
+Unknown command: ${command}`))
   cli.showHelp(1)
 }
 
@@ -51,17 +64,14 @@ Command not implemented: ${command}`))
   cli.showHelp(1)
 }
 
-commands[command](cli.input.slice(1), cli.flags, { cwd: process.cwd(), blinkMobileIdentity })
+commands[command](cli.input.slice(1), cli.flags)
   .catch((err) => {
-    console.log(`There was a problem executing '${command}':
+    console.log(`
+There was a problem executing '${command}':
 
-${err}
+${chalk.red(cli.flags.debug && err && err.stack ? err.stack : err)}
 
 Please fix the error and try again.
-
-`
-)
+`)
     process.exitCode = 1
   })
-  // remove the blow when https://blinkmobile.atlassian.net/browse/CC-22 is done
-  .then(() => process.listenerCount('SIGINT') > 0 ? process.exit() : true)